+++ modules/cpu/fb/LocalFB.h
@@ -82,6 +82,10 @@ struct OSPRAY_SDK_INTERFACE LocalFrameBuffer
   containers::AlignedVector<vec3f> normalBuffer;
   // accumulated, one RGB per pixel
   containers::AlignedVector<vec3f> albedoBuffer;
+  // primitive ID, object ID, and instance ID
+  containers::AlignedVector<uint32> primitiveIDBuffer;
+  containers::AlignedVector<uint32> objectIDBuffer;
+  containers::AlignedVector<uint32> instanceIDBuffer;
 
  protected:
   vec2i getTaskStartPos(const uint32_t taskID) const;