+++ modules/cpu/volume/VolumetricModel.cpp
@@ -72,6 +72,7 @@ void VolumetricModel::commit()
   getSh()->densityScale = getParam<float>("densityScale", 1.f);
   getSh()->anisotropy = getParam<float>("anisotropy", 0.f);
   getSh()->gradientShadingScale = getParam<float>("gradientShadingScale", 0.f);
+  getSh()->userID = getParam<uint32>("id", RTC_INVALID_GEOMETRY_ID);
 }
 
 RTCGeometry VolumetricModel::embreeGeometryHandle() const